pyproject.toml
[project]
name = "golu-svm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.4",
    "pillow>=12.2.0",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
]
